+++ src/seedcase_soil/example_datapackage.py
@@ -1,13 +1,12 @@
-"""Helpers for creating example Data Package metadata."""
+"""Enum of built-in Data Package examples."""
 
-import json
 from enum import StrEnum
 from importlib.resources import files
-from pathlib import Path
-from typing import Any, cast
+
+from .parse_source import Address
 
 
-class ExampleDatapackageName(StrEnum):
+class Example(StrEnum):
     """Available built-in example datapackage names."""
 
     simple = "simple"
@@ -15,22 +14,10 @@ class ExampleDatapackageName(StrEnum):
     flora_imperfect = "flora-imperfect"
     woolly = "woolly"
 
-
-def read_example_datapackage(
-    name: ExampleDatapackageName = ExampleDatapackageName.simple,
-) -> dict[str, Any]:
-    """Return an example datapackage dictionary."""
-    datapackage_path = files("seedcase_soil").joinpath(
-        f"example-datapackages/{name.value}.json"
-    )
-    return cast(dict[str, Any], json.loads(datapackage_path.read_text()))
-
-
-def write_example_datapackage(
-    directory: Path,
-    name: ExampleDatapackageName = ExampleDatapackageName.simple,
-) -> Path:
-    """Write an example datapackage to a directory and return its path."""
-    file_path = directory / "datapackage.json"
-    file_path.write_text(json.dumps(read_example_datapackage(name=name)))
-    return file_path
+    @property
+    def address(self) -> Address:
+        """Return this example's local `Address`."""
+        datapackage_path = files("seedcase_soil").joinpath(
+            f"example-datapackages/{self.value}.json"
+        )
+        return Address(value=str(datapackage_path), local=True)